+++ src/swiss-team.ts
@@ -9,8 +9,12 @@ import {
   typeAColorPreference,
 } from './utilities.js';
 
-import type { PairOptions } from './trace.js';
-import type { CompletedRound, Pairings, Player } from './types.js';
+import type {
+  CompletedRound,
+  PairingOptions,
+  Pairings,
+  Player,
+} from './types.js';
 import type { PlayerState } from './utilities.js';
 
 /**
@@ -137,7 +141,7 @@ function makeAllocateTeamColors(
 function pair(
   players: Player[],
   rounds: CompletedRound[],
-  options?: PairOptions,
+  options?: PairingOptions,
 ): Pairings {
   if (players.length < 2) {
     throw new RangeError('at least 2 players are required');